chore: rename web -> indexdb chore: export MemoryGateway for tests
