+++ src/runtime/gateways/index.ts
@@ -1,3 +1,4 @@
 export * from "./fp-envelope-serialize.js";
 export * as file from "./file/index.js";
+export * as memory from "./memory/gateway.js";
 export * from "./def-serde-gateway.js";
