+++ src/runtime/gateways/indexdb/key-bag-indexdb.ts
@@ -1,6 +1,6 @@
 import { IDBPDatabase, openDB } from "idb";
 import { Logger, ResolveOnce, URI } from "@adviser/cement";
-import { SuperThis, rt } from "@fireproof/core";
+import { type SuperThis, rt } from "@fireproof/core";
 
 export class KeyBagProviderIndexDB implements rt.kb.KeyBagProvider {
   readonly _db: ResolveOnce<IDBPDatabase<unknown>> = new ResolveOnce<IDBPDatabase<unknown>>();
