+++ tests/blockstore/keyed-crypto-indexdb-file.test.ts
@@ -2,7 +2,7 @@ import { bs, PARAM, rt } from "@fireproof/core";
 import { runtimeFn, toCryptoRuntime, URI } from "@adviser/cement";
 import { base58btc } from "multiformats/bases/base58";
 import { mockSuperThis } from "../helpers.js";
-import { KeyBagProviderIndexDB } from "@fireproof/core/web";
+import { KeyBagProviderIndexDB } from "@fireproof/core/indexdb";
 
 describe("KeyBag indexdb and file", () => {
   let url: URI;
